+++ p2panda-stream/src/stream/partial_order.rs
